# open-api-spec/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Anubis",
    "description": "Anubis is a flexible Policy Enforcement solution that makes easier to reuse security policies across different services, assuming the policies entail the same resource.",
    "version": "0.5.0-dev"
  },
  "paths": {
    "/v1/tenants/service_paths": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "List all Service Paths",
        "operationId": "read_service_paths_v1_tenants_service_paths_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Skip",
              "type": "integer",
              "default": 0
            },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Service Paths V1 Tenants Service Paths Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicePath"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenants/": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "List all Tenants",
        "operationId": "read_tenants_v1_tenants__get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Skip",
              "type": "integer",
              "default": 0
            },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Tenants V1 Tenants  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tenant"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "services"
        ],
        "summary": "Create a new Tenant",
        "operationId": "create_tenant_v1_tenants__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ]
      }
    },
    "/v1/tenants/{tenant_id}": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "Get a Tenant",
        "operationId": "read_tenant_v1_tenants__tenant_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            },
            "name": "tenant_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "services"
        ],
        "summary": "Delete a Tenant",
        "operationId": "delete_service_v1_tenants__tenant_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            },
            "name": "tenant_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenants/{tenant_id}/service_paths": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "List Service Paths inside a Tenant",
        "operationId": "read_tenant_service_paths_v1_tenants__tenant_id__service_paths_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            },
            "name": "tenant_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Name",
              "type": "string"
            },
            "name": "name",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Skip",
              "type": "integer",
              "default": 0
            },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Tenant Service Paths V1 Tenants  Tenant Id  Service Paths Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicePath"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "services"
        ],
        "summary": "Create a new Service Path inside a Tenant",
        "operationId": "create_service_path_v1_tenants__tenant_id__service_paths_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            },
            "name": "tenant_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicePathCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenants/{tenant_id}/service_paths/{service_path_id}": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "Get a Service Path inside a Tenant",
        "operationId": "read_service_path_v1_tenants__tenant_id__service_paths__service_path_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            },
            "name": "tenant_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Service Path Id",
              "type": "string"
            },
            "name": "service_path_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicePath"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "services"
        ],
        "summary": "Delete a Service Path inside a Tenant",
        "operationId": "delete_service_path_v1_tenants__tenant_id__service_paths__service_path_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            },
            "name": "tenant_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Service Path Id",
              "type": "string"
            },
            "name": "service_path_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ServicePath": {
        "title": "ServicePath",
        "required": [
          "path",
          "id",
          "tenant_id"
        ],
        "type": "object",
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "parent_id": {
            "title": "Parent Id",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "children": {
            "title": "Children",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServicePath"
            },
            "default": []
          }
        }
      },
      "ServicePathCreate": {
        "title": "ServicePathCreate",
        "required": [
          "path"
        ],
        "type": "object",
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          }
        }
      },
      "Tenant": {
        "title": "Tenant",
        "required": [
          "name",
          "id"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "service_paths": {
            "title": "Service Paths",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServicePath"
            },
            "default": []
          }
        }
      },
      "TenantCreate": {
        "title": "TenantCreate",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "OptionalHTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}